Don't remove prefixed "/" on absolute paths
When using the generator from a path that is not a parent of the file name, the absolute path prefix "/" gets removed on Linux in processPreprocessorLine for scoped files. In custom code, it now assumes a relative path instead of a full path and thus I cannot find the file.